+++ Leibit.Client.WPF/ViewModels/MainViewModel.cs
@@ -6,12 +6,12 @@ using Leibit.Client.WPF.Windows.About.ViewModels;
 using Leibit.Client.WPF.Windows.About.Views;
 using Leibit.Client.WPF.Windows.DelayJustification.ViewModels;
 using Leibit.Client.WPF.Windows.DelayJustification.Views;
+using Leibit.Client.WPF.Windows.Display.ViewModels;
+using Leibit.Client.WPF.Windows.Display.Views;
 using Leibit.Client.WPF.Windows.ESTWSelection.ViewModels;
 using Leibit.Client.WPF.Windows.ESTWSelection.Views;
 using Leibit.Client.WPF.Windows.LocalOrders.ViewModels;
 using Leibit.Client.WPF.Windows.LocalOrders.Views;
-using Leibit.Client.WPF.Windows.PlatformDisplay.ViewModels;
-using Leibit.Client.WPF.Windows.PlatformDisplay.Views;
 using Leibit.Client.WPF.Windows.Settings.ViewModels;
 using Leibit.Client.WPF.Windows.Settings.Views;
 using Leibit.Client.WPF.Windows.SystemState.ViewModels;
@@ -74,7 +74,7 @@ namespace Leibit.Client.WPF.ViewModels
         private CommandHandler m_EstwSelectionCommand;
         private CommandHandler m_TrainProgressInformationCommand;
         private CommandHandler m_TimeTableCommand;
-        private CommandHandler m_PlatformDisplayCommand;
+        private CommandHandler m_DisplayCommand;
         private CommandHandler m_TrainScheduleCommand;
         private CommandHandler m_SystemStateCommand;
         private CommandHandler m_SaveLayoutCommand;
@@ -98,7 +98,7 @@ namespace Leibit.Client.WPF.ViewModels
             m_EstwSelectionCommand = new CommandHandler(__ShowEstwSelectionWindow, false);
             m_TrainProgressInformationCommand = new CommandHandler(__ShowTrainProgressInformationWindow, false);
             m_TimeTableCommand = new CommandHandler<Station>(__ShowTimeTableWindow, true);
-            m_PlatformDisplayCommand = new CommandHandler(__ShowPlatformDisplay, false);
+            m_DisplayCommand = new CommandHandler(__ShowDisplay, false);
             m_TrainScheduleCommand = new CommandHandler(__ShowTrainScheduleWindow, false);
             m_SystemStateCommand = new CommandHandler(__ShowSystemStateWindow, false);
             m_SaveLayoutCommand = new CommandHandler(__SaveLayout, true);
@@ -418,12 +418,12 @@ namespace Leibit.Client.WPF.ViewModels
         }
         #endregion
 
-        #region [PlatformDisplayCommand]
-        public ICommand PlatformDisplayCommand
+        #region [DisplayCommand]
+        public ICommand DisplayCommand
         {
             get
             {
-                return m_PlatformDisplayCommand;
+                return m_DisplayCommand;
             }
         }
         #endregion
@@ -761,14 +761,14 @@ namespace Leibit.Client.WPF.ViewModels
                             else
                                 continue;
 
-                        case eChildWindowType.PlatformDisplay:
+                        case eChildWindowType.Display:
                             var parts = SerializedWindow.Tag?.ToString().Split(';');
 
                             if (parts != null && parts.Length >= 3 && Enum.TryParse(parts[0], out eDisplayType type))
                             {
-                                Window = new PlatformDisplayView();
+                                Window = new DisplayView();
 
-                                var vm = new PlatformDisplayViewModel(Window.Dispatcher, m_CurrentArea);
+                                var vm = new DisplayViewModel(Window.Dispatcher, m_CurrentArea);
                                 vm.SelectedDisplayType = vm.DisplayTypes.FirstOrDefault(x => x.Type == type);
                                 vm.SelectedStation = vm.StationList.FirstOrDefault(s => s.ShortSymbol == parts[1]);
                                 vm.SelectedTrack = vm.TrackList.FirstOrDefault(t => t.Name == parts[2]);
@@ -853,11 +853,11 @@ namespace Leibit.Client.WPF.ViewModels
                     var vm = Window.DataContext as TrainCompositionViewModel;
                     SerializedWindow.Tag = vm.TrainNumber;
                 }
-                else if (Window is PlatformDisplayView)
+                else if (Window is DisplayView)
                 {
-                    SerializedWindow.Type = eChildWindowType.PlatformDisplay;
+                    SerializedWindow.Type = eChildWindowType.Display;
 
-                    var vm = Window.DataContext as PlatformDisplayViewModel;
+                    var vm = Window.DataContext as DisplayViewModel;
                     SerializedWindow.Tag = $"{(int)vm.SelectedDisplayType.Type};{vm.SelectedStation?.ShortSymbol};{vm.SelectedTrack?.Name}";
                 }
                 else
@@ -983,11 +983,11 @@ namespace Leibit.Client.WPF.ViewModels
         }
         #endregion
 
-        #region [__ShowPlatformDisplay]
-        private void __ShowPlatformDisplay()
+        #region [__ShowDisplay]
+        private void __ShowDisplay()
         {
-            var window = new PlatformDisplayView();
-            var vm = new PlatformDisplayViewModel(window.Dispatcher, m_CurrentArea);
+            var window = new DisplayView();
+            var vm = new DisplayViewModel(window.Dispatcher, m_CurrentArea);
 
             __OpenChildWindow(window, vm);
         }
@@ -1134,7 +1134,7 @@ namespace Leibit.Client.WPF.ViewModels
             m_SaveAsCommand.SetCanExecute(true);
             m_EstwSelectionCommand.SetCanExecute(true);
             m_TrainProgressInformationCommand.SetCanExecute(true);
-            m_PlatformDisplayCommand.SetCanExecute(true);
+            m_DisplayCommand.SetCanExecute(true);
             m_SystemStateCommand.SetCanExecute(true);
             IsTrainScheduleEnabled = true;
