+++ Leibit.Client.WPF/Windows/Display/Views/DisplayView.xaml.cs
@@ -1,0 +1,16 @@
+﻿using Leibit.Controls;
+
+namespace Leibit.Client.WPF.Windows.Display.Views
+{
+    /// <summary>
+    /// Interaction logic for DisplayView.xaml
+    /// </summary>
+    public partial class DisplayView : ChildWindow
+    {
+        public DisplayView()
+            : base(null)
+        {
+            InitializeComponent();
+        }
+    }
+}
