+++ Leibit.Client.WPF/Windows/Display/Views/PassengerInformation.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Leibit.Client.WPF.Windows.Display.Views
+{
+    /// <summary>
+    /// Interaction logic for PassengerInformation.xaml
+    /// </summary>
+    public partial class PassengerInformation : UserControl
+    {
+        public PassengerInformation()
+        {
+            InitializeComponent();
+        }
+    }
+}
