+++ Leibit.Client.WPF/Windows/Display/Views/PlatformDisplayLarge.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Leibit.Client.WPF.Windows.Display.Views
+{
+    /// <summary>
+    /// Interaction logic for PlatformDisplayLarge.xaml
+    /// </summary>
+    public partial class PlatformDisplayLarge : UserControl
+    {
+        public PlatformDisplayLarge()
+        {
+            InitializeComponent();
+        }
+    }
+}
