+++ Leibit.Client.WPF/Windows/Display/Views/PlatformDisplaySmall.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Leibit.Client.WPF.Windows.Display.Views
+{
+    /// <summary>
+    /// Interaction logic for PlatformDisplaySmall.xaml
+    /// </summary>
+    public partial class PlatformDisplaySmall : UserControl
+    {
+        public PlatformDisplaySmall()
+        {
+            InitializeComponent();
+        }
+    }
+}
