--- Leibit.Client.WPF/Windows/PlatformDisplay/Views/PlatformDisplayLarge.xaml.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace Leibit.Client.WPF.Windows.PlatformDisplay.Views
-{
-    /// <summary>
-    /// Interaction logic for PlatformDisplayLarge.xaml
-    /// </summary>
-    public partial class PlatformDisplayLarge : UserControl
-    {
-        public PlatformDisplayLarge()
-        {
-            InitializeComponent();
-        }
-    }
-}
