--- Leibit.Client.WPF/Windows/PlatformDisplay/Views/PlatformDisplayView.xaml.cs
@@ -1,16 +1,0 @@
-﻿using Leibit.Controls;
-
-namespace Leibit.Client.WPF.Windows.PlatformDisplay.Views
-{
-    /// <summary>
-    /// Interaction logic for PlatformDisplayView.xaml
-    /// </summary>
-    public partial class PlatformDisplayView : ChildWindow
-    {
-        public PlatformDisplayView()
-            : base(null)
-        {
-            InitializeComponent();
-        }
-    }
-}
