+++ Leibit.Entities/Enums.cs
@@ -53,7 +53,7 @@ namespace Leibit.Entities
         LocalOrders = 7,
         SystemState = 8,
         TrainComposition = 9,
-        PlatformDisplay = 10,
+        Display = 10,
     }
 
     public enum eTrainState
